+++ sdk/resourcemanager/nginx/armnginx/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Release History
 
+## 2.0.0 (2022-10-13)
+### Breaking Changes
+
+- Function `*CertificatesClient.BeginCreate` has been removed
+- Function `*DeploymentsClient.BeginCreate` has been removed
+- Struct `CertificatesClientBeginCreateOptions` has been removed
+- Struct `CertificatesClientCreateResponse` has been removed
+- Struct `DeploymentsClientBeginCreateOptions` has been removed
+- Struct `DeploymentsClientCreateResponse` has been removed
+
+### Features Added
+
+- New function `*CertificatesClient.BeginCreateOrUpdate(context.Context, string, string, string, *CertificatesClientBeginCreateOrUpdateOptions) (*runtime.Poller[CertificatesClientCreateOrUpdateResponse], error)`
+- New function `*DeploymentsClient.BeginCreateOrUpdate(context.Context, string, string, *DeploymentsClientBeginCreateOrUpdateOptions) (*runtime.Poller[DeploymentsClientCreateOrUpdateResponse], error)`
+- New struct `CertificatesClientBeginCreateOrUpdateOptions`
+- New struct `CertificatesClientCreateOrUpdateResponse`
+- New struct `DeploymentsClientBeginCreateOrUpdateOptions`
+- New struct `DeploymentsClientCreateOrUpdateResponse`
+- New field `ProtectedFiles` in struct `ConfigurationProperties`
+
+
 ## 1.0.1 (2022-10-12)
 ### Other Changes
 - Loosen Go version requirement.